remove recognitionPoints and use geometry storage

+++ include/utils/StrokeRecognition.h
@@ -71,7 +71,7 @@ struct StrokeScore {
     int decision = 0;
 };
 
-int stroke_recognition(const QMap<long long, QPointF> &points, StrokeVariables &variables,
+int stroke_recognition(const QMap<qint64, QPointF> &points, StrokeVariables &variables,
                        StrokeResult &result);
 
 #define RECOG_UNKNOWN 0
